apps/ml-pipeline/pyproject.toml
[tool.poetry]
name = "ml-pipeline"
version = "0.1.0"
description = "Operation Obsidian ML Training Pipeline"
authors = ["Pocket Ops Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.0"
pandas = "^2.1.0"
scikit-learn = "^1.3.0"
mlflow = "^2.8.0"
psycopg2-binary = "^2.9.0"
optuna = "^3.4.0"
joblib = "^1.3.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
schedule = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
